[#80] Refactor to follow Room guidelines
