--- app/src/main/java/com/example/wholesomemornings/model/TestActivity.kt
@@ -1,4 +1,0 @@
-package com.example.wholesomemornings.model
-
-class TestActivity(type: String, name: String) : Activity(type, name) {
-}
