+++ app/src/main/java/com/example/wholesomemornings/model/entities/TestActivity.kt
@@ -1,0 +1,6 @@
+package com.example.wholesomemornings.model.entities
+
+import com.example.wholesomemornings.model.entities.Activity
+
+class TestActivity(type: String, name: String) : Activity(type, name) {
+}
