+++ app/src/main/java/com/example/wholesomemornings/model/repositories/ActivityRepository.kt
@@ -1,7 +1,5 @@
 package com.example.wholesomemornings.model.repositories
 
-import com.example.wholesomemornings.model.Activity
-
 class ActivityRepository {
 
     fun getActivity(id: Int) {
