+++ app/src/main/java/com/example/wholesomemornings/view/adapter/ClickableActivityListener.kt
@@ -1,6 +1,6 @@
 package com.example.wholesomemornings.view.adapter
 
-import com.example.wholesomemornings.model.Activity
+import com.example.wholesomemornings.model.entities.Activity
 
 interface ClickableActivityListener {
     fun onActivityClicked(activity: Activity, position: Int)
