+++ app/src/main/java/com/example/wholesomemornings/view/fragment/ActivityFragment.kt
@@ -4,15 +4,12 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
-import androidx.core.os.bundleOf
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.wholesomemornings.R
-import com.example.wholesomemornings.model.Activity
-import com.example.wholesomemornings.model.TestActivity
+import com.example.wholesomemornings.model.entities.Activity
+import com.example.wholesomemornings.model.entities.TestActivity
 import com.example.wholesomemornings.view.adapter.ActivitiesAdapter
 import com.example.wholesomemornings.view.adapter.ClickableActivityListener
 import com.example.wholesomemornings.viewmodel.ActivitiesViewmodel
@@ -35,7 +32,12 @@ class ActivityFragment : Fragment(), ClickableActivityListener {
         super.onViewCreated(view, savedInstanceState)
 
         activitiesAdapter = ActivitiesAdapter(this)
-        activitiesAdapter.activities.addAll(listOf(TestActivity("test", "Test Activity")))
+        activitiesAdapter.activities.addAll(listOf(
+            TestActivity(
+                "test",
+                "Test Activity"
+            )
+        ))
 
         viewModel = ViewModelProviders.of(this).get(ActivitiesViewmodel::class.java)
         viewModel.refresh()
